@@ -1,14 +1,27 @@
-import React, { useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
+import { ThemeProvider } from 'styled-components';
 import Routes from './Routes';
-import { GlobalStyle, Page } from './styles';
+import NavBar from './components/App/NavBar';
+import { AppBox, GlobalStyle } from './styles';
+import ToggleTheme from './components/App/ToggleTheme';
+import { ThemeContext } from './contexts/ThemeContext';
+import BoxLoader from './components/App/BoxLoader';
 
-const App = (): JSX.Element => {
+function App(): JSX.Element {
+  const { theme } = useContext(ThemeContext);
+  const [pageLoaded, setPageLoaded] = useState(false);
+
   return (
-    <Page>
-      <GlobalStyle />
-      <Routes />
-    </Page>
+    <AppBox onLoad={() => setPageLoaded(true)}>
+      <ThemeProvider theme={theme}>
+        <GlobalStyle />
+        <ToggleTheme />
+        {!pageLoaded && <BoxLoader />}
+        <NavBar />
+        <Routes />
+      </ThemeProvider>
+    </AppBox>
   );
-};
+}
 
 export default App;
