@@ -1,153 +1,127 @@
-import styled from 'styled-components';
 import { Link } from 'react-router-dom';
+import styled from 'styled-components';
 
 export const NavBarBox = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 55px;
-  width: 100%;
-  padding: 30px 80px;
   position: fixed;
-  top: 0;
-  left: 0;
+  top: 50%;
+  transform: translateY(-50%);
+  left: 30px;
   z-index: 95;
-  background: ${({ theme }) => theme.colors.background};
-  box-shadow: 0px 0px 0px rgba(0, 0, 0, 0);
-  transition: all 0.2s linear;
 
-  @media (max-width: 900px) {
-    padding: 30px 40px;
-  }
-  /* background: gray; */
-`;
-
-export const Logo = styled(Link)`
-  font-weight: bold;
-  font-size: 20px;
-  color: ${({ theme }) => theme.colors.white};
-  background: ${({ theme }) => theme.similarColors.primaryColor};
-  min-width: 40px;
-  height: 40px;
-  border-radius: 10px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const MenuIcon = styled.div`
-  width: 35px;
-  height: 35px;
-  border-radius: 10px;
-  position: relative;
-  overflow: hidden;
-  cursor: pointer;
-  display: none;
-  animation: fadeIn 0.2s linear;
-
-  @media (max-width: 900px) {
-    display: unset;
+  @media (max-width: 1150px) {
+    background: ${({ theme }) => theme.colors.background};
+    height: 60px;
+    top: 0;
+    left: 0;
+    transform: translateY(0);
+    width: 100%;
+    box-shadow: 0px 0px 8px ${({ theme }) => theme.colors.boxShadow2};
   }
 `;
-export const MenuIconBar = styled.div`
-  width: 25px;
-  height: 3px;
-  background: ${({ theme }) => theme.similarColors.primaryColor};
-  position: absolute;
-  left: 4px;
-  transition: all 0.2s linear;
-
-  &:nth-child(1) {
-    top: 5px;
-  }
-
-  &:nth-child(2) {
-    top: 14px;
-  }
-
-  &:nth-child(3) {
-    top: 23px;
-  }
-`;
-
-export const ToggleMenuCheckBox = styled.input`
-  display: none;
-
-  &:checked ~ .menu-icon div:nth-child(1) {
-    transform: rotateZ(45deg);
-    top: 13px;
-    left: 2px;
-  }
-
-  &:checked ~ .menu-icon div:nth-child(2) {
-    left: 100%;
-  }
-
-  &:checked ~ .menu-icon div:nth-child(3) {
-    transform: rotateZ(-45deg);
-    top: 13px;
-    left: 2px;
-  }
-
-  &:checked ~ .nav-bar-actions {
-    display: flex;
-  }
-`;
-
 export const NavBarActions = styled.div`
   display: flex;
-  flex-wrap: wrap;
-  position: relative;
+  flex-direction: column;
 
-  @media (max-width: 900px) {
-    flex-direction: column;
-    justify-content: space-around;
-    position: fixed;
-    top: 60px;
-    left: 0;
-    height: calc(100vh - 60px);
+  @media (max-width: 1150px) {
+    align-items: center;
+    justify-content: center;
     width: 100%;
-    padding: 40px 0;
+    height: 100vh;
     background: ${({ theme }) => theme.colors.background};
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: 96;
     display: none;
     animation: fadeIn 0.2s linear;
   }
 `;
-
 export const NavBarAction = styled(Link)`
-  font-size: 18px;
-  color: ${({ theme }) => theme.colors.textColor};
-  text-align: center;
-  transition: 0.2s linear;
-  margin: 0 15px;
-  padding: 2px 0;
-  position: relative;
+  height: 50px;
+  width: 50px;
+  max-width: 50px;
+  display: flex;
+  align-items: center;
+  background: ${({ theme }) => theme.colors.background};
+  border-radius: 10px;
+  margin-bottom: 40px;
+  overflow: hidden;
+  box-shadow: inset -3px -3px 10px ${({ theme }) => theme.colors.boxShadow1},
+    inset 8px 8px 25px ${({ theme }) => theme.colors.boxShadow2};
   transition: all 0.2s linear;
 
-  @media (max-width: 900px) {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 8vh;
-    /* border: 3px solid ${({ theme }) => theme.colors.textColor}; */
+  &:last-of-type {
+    margin-bottom: 0;
+  }
+
+  &.selected {
+    background: ${({ theme }) => theme.colors.background};
+    box-shadow: 3px 3px 10px ${({ theme }) => theme.colors.boxShadow2},
+      -2px -2px 10px ${({ theme }) => theme.colors.boxShadow1};
     border-radius: 10px;
-    margin: 0 40px;
   }
 
-  /* &::before {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    right: 0;
-    width: 0%;
-    height: 2px;
-    background: ${({ theme }) => theme.similarColors.primaryColor};
-    transition: all 0.2s linear;
+  @media (min-width: 1150px) {
+    &:hover {
+      width: 100%;
+      max-width: 100%;
+    }
   }
 
-  &.selected-link::before {
-    left: 0;
-    width: 100%;
-  } */
+  @media (max-width: 1150px) {
+    width: 60vw;
+    max-width: unset;
+  }
+
+  @media (max-width: 600px) {
+    width: calc(100vw - 80px);
+    max-width: unset;
+  }
+`;
+export const NavBarActionLabel = styled.p`
+  margin-right: 13px;
+  font-size: 16px;
+
+  @media (max-width: 900px) {
+    font-size: 1rem;
+  }
+`;
+export const NavBarActionIcon = styled.span`
+  font-size: 30px;
+  margin: 0 10px;
+
+  @media (max-width: 900px) {
+    font-size: 1.4rem;
+  }
+`;
+export const ToggleNavBar = styled.input`
+  display: none;
+
+  &:checked + div {
+    display: flex;
+  }
+`;
+export const ToggleNavBarLabel = styled.label`
+  top: 10px;
+  right: 20px;
+  position: absolute;
+  width: 40px;
+  height: 40px;
+  display: none;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+
+  &:hover {
+    background: ${({ theme }) => theme.colors.hoverColor};
+    border-radius: 5px;
+  }
+
+  @media (max-width: 1150px) {
+    display: flex;
+  }
+`;
+export const ToggleNavBarIcon = styled.span`
+  font-size: 1.7rem;
+  color: ${({ theme }) => theme.colors.secondary};
 `;
