@@ -2,24 +2,61 @@ import styled from 'styled-components';
 
 export const SkillBox = styled.div`
   display: flex;
+  align-items: center;
+  border: 2px solid ${({ color }) => color};
+  border-radius: 10px;
+  padding: 10px;
+  width: 300px;
+  margin-bottom: 20px;
+  cursor: pointer;
+
+  &:last-of-type {
+    margin-bottom: 0;
+  }
+
+  &:hover > div > p {
+    transform: translateY(0);
+  }
+
+  @media (max-width: 1150px) {
+    width: 60vw;
+  }
+
+  @media (max-width: 600px) {
+    width: calc(100vw - 80px);
+  }
+`;
+export const SkillIcon = styled.img`
+  margin-right: 10px;
+  width: 43px;
+
+  @media (max-width: 900px) {
+    width: 30px;
+  }
+`;
+export const SkillInfo = styled.div`
+  display: flex;
   flex-direction: column;
   width: 100%;
-  margin-bottom: 10px;
-  color: ${({ theme }) => theme.similarColors.primaryColor};
+  transition: all 0.2s linear;
 `;
 export const SkillName = styled.p`
-  margin-bottom: 5px;
+  color: ${({ color }) => color};
   font-weight: bold;
+  text-transform: uppercase;
+  font-size: 1rem;
+  margin-bottom: 2px;
+  transform: translateY(4px);
+  transition: all 0.2s linear;
+
+  @media (max-width: 600px) {
+    font-size: 0.9rem;
+  }
 `;
-export const SkillProgress = styled.div`
-  height: 10px;
-  width: 100%;
-  background: ${({ theme }) => theme.colors.textColor};
+export const SkillPercentage = styled.div`
+  background: ${({ color }) => color};
+  height: 5px;
+  width: 0%;
   border-radius: 5px;
-  overflow: hidden;
+  transition: width 0.2s linear, height 0s;
 `;
-export const SkillBar = styled.div`
-  height: 100%;
-  background: ${({ theme }) => theme.similarColors.primaryColor};
-`;
-export const Progress = styled.div``;
