@@ -1,12 +1,21 @@
 import React from 'react';
-import { IRoute } from '../../routes';
+import { ContactBox, TopWave, BottomWave, Or } from './styles';
 
-const Contact = (props: IRoute): JSX.Element => {
+import topWave from './assets/topWave.svg';
+import bottomWave from './assets/bottomWave.svg';
+import SendEmail from '../../components/Contact/SendEmail';
+import FollowMe from '../../components/Contact/FollowMe';
+
+function Contact(): JSX.Element {
   return (
-    <div>
-      <p>Início</p>
-    </div>
+    <ContactBox>
+      <SendEmail />
+      <Or>or</Or>
+      <FollowMe />
+      <TopWave src={topWave} />
+      <BottomWave src={bottomWave} />
+    </ContactBox>
   );
-};
+}
 
 export default Contact;
