@@ -1,16 +1,22 @@
 import styled from 'styled-components';
+import { PageBox } from '../../styles';
 
-export const ProjectsBox = styled.div`
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  padding: 40px;
-  animation: fadeIn 0.2s linear;
-
-  @media (max-width: 900px) {
-    padding-top: 70px;
+export const ProjectsBox = styled(PageBox)`
+  @media (max-width: 1050px) {
+    padding: 100px 0;
   }
+`;
+export const TopWave = styled.img`
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  width: 100vw;
+  z-index: -1;
+`;
+export const BottomWave = styled.img`
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  width: 100vw;
+  z-index: -2;
 `;
