@@ -1,13 +1,18 @@
 import React from 'react';
-import WorksList from '../../components/Works/WorksList';
-import { WorksBox } from './styles';
+import { WorksBox, TopWave, BottomWave } from './styles';
 
-const Works = (): JSX.Element => {
+import topWave from './assets/topWave.svg';
+import bottomWave from './assets/bottomWave.svg';
+import WorksList from '../../components/Works/WorksList';
+
+function Works(): JSX.Element {
   return (
     <WorksBox>
       <WorksList />
+      <TopWave src={topWave} />
+      <BottomWave src={bottomWave} />
     </WorksBox>
   );
-};
+}
 
 export default Works;
